+++ src/ErrorPlot.tsx
@@ -1,0 +1,7 @@
+// Stub component for ErrorPlot visualization
+function ErrorPlot() {
+  console.log('ErrorPlot rendered');
+  return <div>Error Plot Component</div>;
+}
+
+export { ErrorPlot }
