+++ src/LocalError.tsx
@@ -1,0 +1,6 @@
+// Stub component for LocalError visualization
+function LocalError() {
+  return <div>Local Error Component</div>;
+}
+
+export { LocalError }
