+++ src/utils.ts
@@ -1,0 +1,11 @@
+function nextId(table: { id: number }[]) {
+  return table.sort((a, b) => a.id - b.id).reduce((acc, curr) => {
+    if (acc === curr.id) {
+      return acc + 1;
+    } else {
+      return acc;
+    }
+  }, 0);
+}
+
+export { nextId }
